+++ angular_compiler/CHANGELOG.md
@@ -1,5 +1,7 @@
 * Requires `source_gen ^0.7.4` (was previously `^0.7.0`).
 
+* `linkToReference` now requires a second parameter, a `LibraryReader`.
+
 ## 0.4.0-alpha+4
 
 ### Breaking changes